[FRONT-015] error handling when dateto is smaller than dateFrom
If event start date and end date is the same, and the user pick a start time that is after end time,
update the endtime to be the same as start time.

If the user choose to pick a end time/date that is before the start time/date, an error message is shown.